Now can download audio only. TODO: Implement info getter. To list available formats

+++ backend/Domain/Classes/Audio.cs
@@ -5,6 +5,6 @@ namespace MediaDownloader.Domain.Classes;
 public class Audio : AbstractMediaFile
 {
   public required EAudioExtension Extension { get; set; }
-  public required long Bitrate { get; set; }
+  public required float Bitrate { get; set; }
   public required int DurationSec { get; set; }
 }
